+++ requirements-dev.lock
@@ -1,0 +1,83 @@
+# generated by rye
+# use `rye lock` or `rye sync` to update this lockfile
+#
+# last locked with the following flags:
+#   pre: false
+#   features: []
+#   all-features: false
+#   with-sources: false
+
+-e file:.
+annotated-types==0.6.0
+    # via pydantic
+anyio==4.3.0
+    # via starlette
+cachetools==5.3.3
+    # via google-auth
+certifi==2024.2.2
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via uvicorn
+fastapi==0.110.1
+    # via droras
+google-api-core==2.18.0
+    # via google-cloud-core
+    # via google-cloud-firestore
+google-auth==2.29.0
+    # via google-api-core
+    # via google-cloud-core
+    # via google-cloud-firestore
+google-cloud-core==2.4.1
+    # via google-cloud-firestore
+google-cloud-firestore==2.16.0
+    # via droras
+googleapis-common-protos==1.63.0
+    # via google-api-core
+    # via grpcio-status
+grpcio==1.62.1
+    # via google-api-core
+    # via grpcio-status
+grpcio-status==1.62.1
+    # via google-api-core
+h11==0.14.0
+    # via uvicorn
+idna==3.7
+    # via anyio
+    # via requests
+proto-plus==1.23.0
+    # via google-api-core
+    # via google-cloud-firestore
+protobuf==4.25.3
+    # via google-api-core
+    # via google-cloud-firestore
+    # via googleapis-common-protos
+    # via grpcio-status
+    # via proto-plus
+pyasn1==0.6.0
+    # via pyasn1-modules
+    # via rsa
+pyasn1-modules==0.4.0
+    # via google-auth
+pydantic==2.7.0
+    # via fastapi
+pydantic-core==2.18.1
+    # via pydantic
+requests==2.31.0
+    # via droras
+    # via google-api-core
+rsa==4.9
+    # via google-auth
+sniffio==1.3.1
+    # via anyio
+starlette==0.37.2
+    # via fastapi
+typing-extensions==4.11.0
+    # via fastapi
+    # via pydantic
+    # via pydantic-core
+urllib3==2.2.1
+    # via requests
+uvicorn==0.29.0
+    # via droras
