+++ src/droras/__init__.py
@@ -1,0 +1,12 @@
+import logging
+import os
+
+from . import config
+
+# logging
+logging.basicConfig(
+    level=logging.DEBUG,
+    format="%(asctime)s.%(msecs)03d - %(name)s - %(levelname)s - %(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
+    handlers=[logging.FileHandler(os.path.join(config.LOG_DIR, "app.log")), logging.StreamHandler()],
+)
